Fluid conduit crashes game if it has non-existant fluid
If a fluid is removed from the game, minecraft / forge will assign
the fluid to the instance "Fluids.EMPTY".

The nbt fluid string will still be kept (to the non-existant fluid)
initially, but if you lookup the fluid string in the registries,
it will return "Fluids.EMPTY".

The fix is to set the lockedFluid to null on deserialization,
if the fluid, after lookup, is equal to "Fluids.EMPTY".

closes #382