AE2 conduits crashes server on chunkload
If you try to create a IManagedGridNode inside an unloaded chunk,
AE2 will throw an IllegalStateException.

The documentation for IManagedGridNode.create() is clear:
"This should only be called when the node is in a ticking chunk".
The suggested solution is GridHelper.onFirstTick().

The fix: When calling ConduitBlockEntity.onload(), make sure at least
1 server tick has happened by calling checkConnection.isInitialized().

closes #434