Return to classic versioning scheme. More in description.
During the release of v6.0.9-alpha, the Minecraft version was added to the version string. This quickly changed however to appending it to the jar name instead, as it was breaking some versioning systems behind the scenes.
This workaround is no longer required now that we're launching a new release series.