chore: Mark EnderBlockEntity dataslots as deprecated.
These are going to be completely removed.

Blocks should either move syncing into the menu if the data is only used by the menu;
into the update tag (and fire relevant block update events) if it is required by the block and is updated infrequently;
or into a custom packet and sent when required for complex/fast changing data (do so sparingly!)

I will slowly begin converting things to use this new structure.