feat: More progress on ultimine compatibility
This implementation looks about finished, although the inability to break the shape that ultimine destroys results in unpredictable behaviour such as destroying all fluid conduits that are touching any bundles, or any facades of the same type across an entire run of bundles.

I've reached back out to FTB to see if we can improve this further.